+++ tokenizers/src/models/wordlevel/mod.rs
@@ -208,7 +208,7 @@ impl Model for WordLevel {
         let mut vocab_file = File::create(&vocab_path)?;
         let order_vocab_iter = OrderedVocabIter::new(&self.vocab_r);
         let serialized = serde_json::to_string(&order_vocab_iter)?;
-        vocab_file.write_all(&serialized.as_bytes())?;
+        vocab_file.write_all(serialized.as_bytes())?;
 
         Ok(vec![vocab_path])
     }